+++ NEWS.md
@@ -3,6 +3,7 @@
 - Ability to add multiple `e_title`.
 - `draw` argument added to `e_charts`, to set whether to draw the chart, see bullet below.
 - `e_draw_p` added to choose draw chart.
+- `e_inspect` to inspect the output of echarts4r.
 
 # echarts4r 0.2.1