+++ NEWS.md
@@ -4,6 +4,7 @@
 - `draw` argument added to `e_charts`, to set whether to draw the chart, see bullet below.
 - `e_draw_p` added to choose draw chart.
 - `e_inspect` to inspect the output of echarts4r.
+- `echarts_from_json` to create an echarts4r object from JSON options.
 
 # echarts4r 0.2.1