+++ NEWS.md
@@ -6,6 +6,7 @@
 - `e_inspect` to inspect the output of echarts4r.
 - `echarts_from_json` to create an echarts4r object from JSON options.
 - `e_axis` family of functions (e.g.: `e_x_axis`) now accepts a column name as first argument to easily adapt the range of the axes, in a similar fashion as with `e_visual_map`.
+- `e_charts` family forces data.frame conversion to avoid jsonlite bug.
 
 # echarts4r 0.2.1