+++ NEWS.md
@@ -1,5 +1,9 @@
 # echarts4r 0.2.2
 
+- Bar chart's `bind` argument has been fixed.
+
+# echarts4r 0.2.2
+
 - Ability to add multiple `e_title`.
 - `draw` argument added to `e_charts`, to set whether to draw the chart, see bullet below.
 - `e_draw_p` added to choose draw chart.